completed extraction with ExtractorBase class
Now features are correctly extracted on full images or on tiles and saved in h5 database. The database is read by the ImageMatching class in order to pass the features to the matcher. The matching now is working properly on full images. With tiling the order of keypoints on the images is not kept yet (the keypoints in the h5 database must be used to build the colmap database, the matcher must return only the indexes of the matches on each image pair, starting always from the h5 db).


Former-commit-id: 896575f1b8d515ee179cee2ea4952941da7477bf